update: category api 패키지 분리 및 네이밍, 순서 수정
