--- src/main/java/com/newdeal/ledger/categorytag/mapper/CategoryMapper.java
@@ -1,16 +1,0 @@
-package com.newdeal.ledger.categorytag.mapper;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import com.newdeal.ledger.categorytag.dto.CategoryDto;
-
-@Mapper
-public interface CategoryMapper {
-	List<CategoryDto> findAll();
-
-	List<CategoryDto> findAllCategoriesByType(String type);
-
-	List<CategoryDto> findAllSubcategoriesByParentCno(Integer parentCno);
-}
