--- src/main/java/com/newdeal/ledger/categorytag/service/CategoryService.java
@@ -1,13 +1,0 @@
-package com.newdeal.ledger.categorytag.service;
-
-import java.util.List;
-
-import com.newdeal.ledger.categorytag.dto.CategoryDto;
-
-public interface CategoryService {
-
-	List<CategoryDto> selectAll();
-	List<CategoryDto> getCategoriesByType(String type);
-
-	List<CategoryDto> getSubcategories(Integer parentCno);
-}
