--- src/main/java/com/newdeal/ledger/categorytag/service/CategoryServiceImpl.java
@@ -1,31 +1,0 @@
-package com.newdeal.ledger.categorytag.service;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import com.newdeal.ledger.categorytag.dto.CategoryDto;
-import com.newdeal.ledger.categorytag.mapper.CategoryMapper;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class CategoryServiceImpl implements CategoryService {
-	private final CategoryMapper mapper;
-
-	public List<CategoryDto> selectAll() {
-		return mapper.findAll();
-	}
-
-	public List<CategoryDto> getCategoriesByType(String type){
-		List<CategoryDto> allCategoriesByType = mapper.findAllCategoriesByType(type);
-
-		return allCategoriesByType;
-	}
-
-	public List<CategoryDto> getSubcategories(Integer parentCno){
-		return mapper.findAllSubcategoriesByParentCno(parentCno);
-	}
-
-}
