+++ src/main/resources/mybatis/mapper/category-mapper.xml
@@ -3,10 +3,10 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.newdeal.ledger.categorytag.mapper.CategoryMapper">
+<mapper namespace="com.newdeal.ledger.category.mapper.CategoryMapper">
 
     <!-- parent_cno가 null인 카테고리 리스트를 조회하는 쿼리 -->
-    <select id="findAll" resultType="com.newdeal.ledger.categorytag.dto.CategoryDto">
+    <select id="findCategories" resultType="com.newdeal.ledger.category.dto.CategoryDto">
         SELECT cno,
                type,
                name,
@@ -15,7 +15,7 @@
         WHERE parent_cno IS NULL
     </select>
 
-    <select id="findAllCategoriesByType" resultType="com.newdeal.ledger.categorytag.dto.CategoryDto"
+    <select id="findCategoriesByType" resultType="com.newdeal.ledger.category.dto.CategoryDto"
             parameterType="String">
         SELECT cno,
                type,
@@ -27,15 +27,14 @@
 
     </select>
 
-    <select id="findAllSubcategoriesByParentCno" resultType="com.newdeal.ledger.categorytag.dto.CategoryDto"
+    <select id="findSubcategoriesByCategoryId" resultType="com.newdeal.ledger.category.dto.CategoryDto"
             parameterType="Integer">
         SELECT cno,
                type,
                name,
                parent_cno as parentCno
         FROM category
-        WHERE parent_cno = #{parentCno}
-
+        WHERE parent_cno = #{categoryId}
     </select>
 
 </mapper>
